agrego componente para watchlists
