+++ frontend/app-front/app/routes/watchlists.tsx
@@ -1,29 +1,17 @@
 import { HeaderFull } from "components/ui/header-full";
-import { MovieCard } from "components/ui/movie-card";
 import { SidebarNav } from "components/ui/sidebar-nav";
-import { Clock, Film, Pencil, Plus, Star, TrendingUp } from "lucide-react";
-// import { Pagination } from "components/ui/pagination";
-// import { useTranslation } from "react-i18next";
-// import { useFetcher, useLoaderData } from "react-router-dom";
-import dayjs from "dayjs";
-import utc from "dayjs/plugin/utc";
+import { Film, Plus, Star, TrendingUp } from "lucide-react";
 import { useState } from "react";
 import { useLoaderData } from "react-router-dom";
 import { getWatchLists } from "services/api/flixy/server/watchlists";
 import type { ApiResponse } from "../../services/api/flixy/types/overall";
 import type { Route } from "./+types/movies";
 
-import relativeTime from "dayjs/plugin/relativeTime";
-
-import { AddMovieWatchList } from "components/ui/add-movie-watchlist";
 import { Button } from "components/ui/button";
+import WatchList from "components/ui/watchlist";
 import "dayjs/locale/en";
 import "dayjs/locale/es";
-import i18n from "i18n/i18n";
 import { getAccessToken } from "services/api/utils";
-
-dayjs.extend(relativeTime);
-dayjs.extend(utc);
 
 interface Movie {
   id: number;
@@ -41,7 +29,7 @@ interface Movie {
   user_rating: number | null;
 }
 
-interface WatchList {
+interface WatchListFace {
   id: number;
   name: string;
   description: string;
@@ -53,7 +41,7 @@ interface WatchList {
 
 interface WatchLists {
   items: {
-    items: WatchList[];
+    items: WatchListFace[];
   };
 }
 
@@ -67,8 +55,6 @@ export async function loader({ request }: Route.LoaderArgs) {
   let apiResponse: ApiResponse = {};
 
   let watchlists: WatchLists = {} as WatchLists;
-
-  console.log(watchlists);
 
   try {
     watchlists = await getWatchLists(page, DEFAULT_PAGE_SIZE, request);
@@ -91,9 +77,7 @@ export async function loader({ request }: Route.LoaderArgs) {
   }
 }
 
-export default function MovieInsights() {
-  dayjs.locale(i18n.language || "en");
-
+export default function WatchListsPage() {
   const apiResponse: ApiResponse = useLoaderData();
   const [watchlists, setWatchlists] = useState<WatchLists>(
     apiResponse.data || []
@@ -188,121 +172,13 @@ export default function MovieInsights() {
         </div>
 
         {/* Watchlists Content */}
-
         <div className="p-6 pt-0">
           {watchlists.items.items.map((watchlist) => (
-            <div key={watchlist.id} className="mb-12">
-              {/* Watchlist Header */}
-              <div className="flex items-center justify-between mb-4">
-                <div className="flex items-center gap-3">
-                  {/* <span className="text-2xl">{watchlist.icon}</span> */}
-                  <div>
-                    <h2 className="text-2xl font-semibold text-white">
-                      {watchlist.name}
-                    </h2>
-                    <div className="flex items-center gap-4 mt-1">
-                      <span className="text-sm text-gray-400">
-                        {watchlist.movies.length} movies
-                      </span>
-                      <div className="flex items-center gap-1 text-xs text-gray-500">
-                        <Clock className="w-3 h-3" />
-                        <span>{dayjs.utc(watchlist.created_at).fromNow()}</span>
-                      </div>
-                      <div className="flex items-center gap-1 text-xs text-gray-500">
-                        <Pencil className="w-3 h-3" />
-                        <span>{dayjs.utc(watchlist.updated_at).fromNow()}</span>
-                      </div>
-                    </div>
-                  </div>
-                </div>
-                <button
-                  onClick={() => {}}
-                  className="mt-5 justify-end bg-transparent border-none text-violet-400 text-sm font-medium cursor-pointer px-4 py-2 rounded-md transition-all duration-200 hover:bg-slate-800"
-                >
-                  See WatchList
-                </button>
-              </div>
-
-              {/* Divider Line */}
-              <div className="h-px bg-gray-600 mb-6" />
-
-              {/* Movie Posters Grid */}
-              <div className="flex gap-4 overflow-x-auto pb-2">
-                {watchlist.movies.length > 5 ? (
-                  <>
-                    {watchlist.movies.slice(0, 4).map((movie) => (
-                      <MovieCard
-                        key={movie.id}
-                        movie={movie}
-                        styles={{
-                          card_width: 128,
-                          card_height: 160,
-                          title_size: "sm",
-                          title_padding: 1,
-                          show_rate_this_movie: false,
-                          star_component_multiplier: 0.5,
-                          show_details_size: "xs",
-                          show_details_padding: 1,
-                        }}
-                        accessToken={undefined}
-                      />
-                    ))}
-
-                    <div className="flex items-center justify-center w-12 text-gray-400 text-2xl">
-                      ...
-                    </div>
-
-                    <MovieCard
-                      key={watchlist.movies[watchlist.movies.length - 1].id}
-                      movie={watchlist.movies[watchlist.movies.length - 1]}
-                      styles={{
-                        card_width: 128,
-                        card_height: 160,
-                        title_size: "sm",
-                        title_padding: 1,
-                        show_rate_this_movie: false,
-                        star_component_multiplier: 0.5,
-                        show_details_size: "xs",
-                        show_details_padding: 1,
-                      }}
-                      accessToken={undefined}
-                    />
-                  </>
-                ) : (
-                  watchlist.movies.map((movie) => (
-                    <MovieCard
-                      key={movie.id}
-                      movie={movie}
-                      styles={{
-                        card_width: 128,
-                        card_height: 160,
-                        title_size: "sm",
-                        title_padding: 1,
-                        show_rate_this_movie: false,
-                        star_component_multiplier: 0.5,
-                        show_details_size: "xs",
-                        show_details_padding: 1,
-                      }}
-                      accessToken={undefined}
-                    />
-                  ))
-                )}
-                <AddMovieWatchList
-                  accessToken={String(apiResponse.accessToken)}
-                />
-              </div>
-            </div>
+            <WatchList
+              accessToken={String(apiResponse.accessToken)}
+              watchlist={watchlist}
+            />
           ))}
-
-          {/* Empty State Message */}
-          {watchlists.items.items.length === 0 && (
-            <div className="text-center py-16 px-8 text-gray-400">
-              <p className="text-lg mb-2">No watchlists yet</p>
-              <p className="text-sm">
-                Create your first watchlist to get started
-              </p>
-            </div>
-          )}
         </div>
       </div>
     </div>
